Move skeleton data API implementation into its own module
To make the diff of the implementation PR easier to read.
