+++ salesforce_functions/data_api/exceptions.py
@@ -1,0 +1,2 @@
+class DataApiError(Exception):
+    pass
